+++ component/backend/Controller/ControlPanel.php
@@ -23,7 +23,7 @@ class ControlPanel extends Controller
 	{
 		parent::__construct($container, $config);
 
-		$this->predefinedTaskList = ['main', 'updategeoip'];
+		$this->predefinedTaskList = ['main'];
 	}
 
 	/**
@@ -37,43 +37,5 @@ class ControlPanel extends Controller
 			->checkAndFixDatabase()
 			->saveMagicVariables()
 			->deleteUpdateSites();
-	}
-
-	/**
-	 * Update the GeoIP database
-	 */
-	public function updategeoip()
-	{
-		if ($this->csrfProtection)
-		{
-			$this->csrfProtection();
-		}
-
-		// Load the GeoIP library if it's not already loaded
-		if (!class_exists('AkeebaGeoipProvider'))
-		{
-			if (@file_exists(JPATH_PLUGINS . '/system/akgeoip/lib/akgeoip.php'))
-			{
-				if (@include_once JPATH_PLUGINS . '/system/akgeoip/lib/vendor/autoload.php')
-				{
-					@include_once JPATH_PLUGINS . '/system/akgeoip/lib/akgeoip.php';
-				}
-			}
-		}
-
-		$geoip = new \AkeebaGeoipProvider();
-		$result = $geoip->updateDatabase();
-
-		$url = 'index.php?option=com_akeebasubs';
-
-		if ($result === true)
-		{
-			$msg = JText::_('COM_AKEEBASUBS_GEOIP_MSG_DOWNLOADEDGEOIPDATABASE');
-			$this->setRedirect($url, $msg);
-		}
-		else
-		{
-			$this->setRedirect($url, $result, 'error');
-		}
 	}
 }