+++ component/backend/Model/ControlPanel.php
@@ -97,43 +97,6 @@ class ControlPanel extends Model
 	}
 
 	/**
-	 * Does the GeoIP database need update?
-	 *
-	 * @param   integer  $maxAge  The maximum age of the db in days (default: 15)
-	 *
-	 * @return  boolean
-	 */
-	public function GeoIPDBNeedsUpdate($maxAge = 15)
-	{
-		$needsUpdate = false;
-
-		if (!$this->hasGeoIPPlugin())
-		{
-			return $needsUpdate;
-		}
-
-		// Get the modification time of the database file
-		$filePath = JPATH_ROOT . '/plugins/system/akgeoip/db/GeoLite2-Country.mmdb';
-		$modTime = @filemtime($filePath);
-
-		// This is now
-		$now = time();
-
-		// Minimum time difference we want (15 days) in seconds
-		if ($maxAge <= 0)
-		{
-			$maxAge = 15;
-		}
-
-		$threshold = $maxAge * 24 * 3600;
-
-		// Do we need an update?
-		$needsUpdate = ($now - $modTime) > $threshold;
-
-		return $needsUpdate;
-	}
-
-	/**
 	 * Removes the update sites for this extension
 	 *
 	 * @since  7.0.1