+++ component/backend/ViewTemplates/ControlPanel/default.blade.php
@@ -8,18 +8,11 @@
 defined('_JEXEC') or die;
 ?>
 
-{{-- Include external sections. Do note how you can include sub-templates in one order and compile them in a completely
-different order using @yield later on! --}}
 @include('admin:com_akeebasubs/ControlPanel/phpversion')
 @include('admin:com_akeebasubs/ControlPanel/graphs')
 @include('admin:com_akeebasubs/ControlPanel/stats')
-{{-- Note: I don't pass $this->hasGeoIPPlugin and $this->geoIPPluginNeedsUpdate. This demonstrates how Blade
-subtemplates can view their parent's variables automatically. --}}
 @include('admin:com_akeebasubs/ControlPanel/geoip')
 
-{{-- Compile the output. Do note that I don't need to wrap it in a section. Content outside a section is yielded
-immediately. Alternatively I could wrap this in a @section/@show block or even @section/@stop and use @yield to
-render it. --}}
 @yield('phpVersionWarning', '')
 
 @yield('geoip', '')